Updated filter for transaction - UGLY:(
Updated the way we pre-process the transaction name so that its easier to determine what we're looking for. This is currently done really ugly, so it needs to change. But I think first, I want to correctly append to a list of unknown entries, and then also update to use csv tables for the transaction lists. Perhaps keep the same dictionary properties, just read in the csv files into the separate dictionary lists.